Move non-generic functions into preferences file
