+++ data/binding/preference.go
@@ -107,3 +107,30 @@ func (b *prefBoundString) replaceProvider(p fyne.Preferences) {
 	b.get = p.String
 	b.set = p.SetString
 }
+
+func lookupExistingBinding(key string, p fyne.Preferences) (preferenceItem, bool) {
+	binds := prefBinds.getBindings(p)
+	if binds == nil {
+		return nil, false
+	}
+
+	if listen, ok := binds.Load(key); listen != nil && ok {
+		return listen, true
+	}
+
+	return nil, false
+}
+
+type prefBoundItem interface {
+	preferenceItem
+	setKey(string)
+	replaceProvider(fyne.Preferences)
+}
+
+func setupPrefItem(listen prefBoundItem, key string, p fyne.Preferences) preferenceItem {
+	listen.setKey(key)
+	listen.replaceProvider(p)
+	binds := prefBinds.ensurePreferencesAttached(p)
+	binds.Store(key, listen)
+	return listen
+}
